refactor(paths): fold bundled skills resolver into paths.ts, drop legacy/skills/
- resolveBundledSkillsDir + looksLikeSkillsDir helper move from
  src/legacy/skills/bundled-dir.ts into src/paths.ts (it's just path
  resolution, belongs alongside other path helpers)
- bundled-dir.test.ts → src/bundled-skills.test.ts
- src/legacy/skills/ directory deleted (was just re-export + impl)
- workspace.ts import updated
