--- src/legacy/skills/index.ts
@@ -1,1 +1,0 @@
-export { resolveBundledSkillsDir } from "./bundled-dir.js";
